Orchestrator (#187)
* Move bee into orchestration

* Prepare orchestration for interfaces

* Create Node interface

* Move NodeGroupOptions

* Move orchestration types

* Create cluster interface

* Fix imports

* Move ClusterOptions

* Notset

* Delete bee interface

* Cleanup

* Change internal types

* Resolve k8s client inheritence

* Unify imports
