+++ pkg/check/cashout/cashout.go
@@ -46,7 +46,7 @@ type CashoutAction struct {
 }
 
 // Check executes settlements check
-func (c *Check) Run(ctx context.Context, cluster *orchestration.Cluster, opts interface{}) (err error) {
+func (c *Check) Run(ctx context.Context, cluster orchestration.Cluster, opts interface{}) (err error) {
 	o, ok := opts.(Options)
 	if !ok {
 		return fmt.Errorf("invalid options type")
