+++ pkg/check/fileretrieval/fileretrieval.go
@@ -59,7 +59,7 @@ func NewCheck() beekeeper.Action {
 	return &Check{}
 }
 
-func (c *Check) Run(ctx context.Context, cluster *orchestration.Cluster, opts interface{}) (err error) {
+func (c *Check) Run(ctx context.Context, cluster orchestration.Cluster, opts interface{}) (err error) {
 	o, ok := opts.(Options)
 	if !ok {
 		return fmt.Errorf("invalid options type")
@@ -75,7 +75,7 @@ func (c *Check) Run(ctx context.Context, cluster *orchestration.Cluster, opts in
 var errFileRetrieval = errors.New("file retrieval")
 
 // defaultCheck uploads files on cluster and downloads them from the last node in the cluster
-func defaultCheck(ctx context.Context, cluster *orchestration.Cluster, o Options) (err error) {
+func defaultCheck(ctx context.Context, cluster orchestration.Cluster, o Options) (err error) {
 	fmt.Println("running file retrieval")
 
 	rnds := random.PseudoGenerators(o.Seed, o.UploadNodeCount)
@@ -167,7 +167,7 @@ func defaultCheck(ctx context.Context, cluster *orchestration.Cluster, o Options
 }
 
 // fullCheck uploads files on cluster and downloads them from the all nodes in the cluster
-func fullCheck(ctx context.Context, cluster *orchestration.Cluster, o Options) (err error) {
+func fullCheck(ctx context.Context, cluster orchestration.Cluster, o Options) (err error) {
 	fmt.Println("running file retrieval (full mode)")
 
 	rnds := random.PseudoGenerators(o.Seed, o.UploadNodeCount)
