+++ pkg/check/kademlia/kademlia.go
@@ -35,7 +35,7 @@ func NewCheck() beekeeper.Action {
 	return &Check{}
 }
 
-func (c *Check) Run(ctx context.Context, cluster *orchestration.Cluster, opts interface{}) (err error) {
+func (c *Check) Run(ctx context.Context, cluster orchestration.Cluster, opts interface{}) (err error) {
 	fmt.Println("running kademlia")
 	o, ok := opts.(Options)
 	if !ok {
