+++ pkg/check/manifest/manifest.go
@@ -57,7 +57,7 @@ func NewCheck() beekeeper.Action {
 
 var errManifest = errors.New("manifest data mismatch")
 
-func (c *Check) Run(ctx context.Context, cluster *orchestration.Cluster, opts interface{}) (err error) {
+func (c *Check) Run(ctx context.Context, cluster orchestration.Cluster, opts interface{}) (err error) {
 	o, ok := opts.(Options)
 	if !ok {
 		return fmt.Errorf("invalid options type")
