+++ pkg/check/pullsync/pullsync.go
@@ -54,7 +54,7 @@ func NewCheck() beekeeper.Action {
 
 var errPullSync = errors.New("pull sync")
 
-func (c *Check) Run(ctx context.Context, cluster *orchestration.Cluster, opts interface{}) (err error) {
+func (c *Check) Run(ctx context.Context, cluster orchestration.Cluster, opts interface{}) (err error) {
 	o, ok := opts.(Options)
 	if !ok {
 		return fmt.Errorf("invalid options type")
