+++ pkg/check/pushsync/check_lightnode.go
@@ -13,7 +13,7 @@ import (
 )
 
 // checkChunks uploads given chunks on cluster and checks pushsync ability of the cluster
-func checkLightChunks(ctx context.Context, cluster *orchestration.Cluster, o Options) error {
+func checkLightChunks(ctx context.Context, cluster orchestration.Cluster, o Options) error {
 	rnds := random.PseudoGenerators(o.Seed, o.UploadNodeCount)
 	fmt.Printf("seed: %d\n", o.Seed)
