+++ pkg/check/smoke/smoke.go
@@ -49,7 +49,7 @@ func NewCheck() beekeeper.Action {
 }
 
 // Check uploads given chunks on cluster and checks pushsync ability of the cluster
-func (c *Check) Run(ctx context.Context, cluster *orchestration.Cluster, opts interface{}) (err error) {
+func (c *Check) Run(ctx context.Context, cluster orchestration.Cluster, opts interface{}) (err error) {
 	o, ok := opts.(Options)
 	if !ok {
 		return fmt.Errorf("invalid options type")
