+++ pkg/orchestration/node.go
@@ -1,19 +1,30 @@
 package orchestration
 
 import (
+	"context"
+	"errors"
+	"time"
+
 	"github.com/ethersphere/beekeeper/pkg/bee"
-	ob "github.com/ethersphere/beekeeper/pkg/orchestration/bee"
+	"github.com/ethersphere/beekeeper/pkg/k8s"
 )
 
-// Node represents Bee node
-type Node struct {
-	name         string
-	clefKey      string
-	clefPassword string
-	client       *bee.Client
-	config       *ob.Config
-	libP2PKey    string
-	swarmKey     string
+// ErrNotSet represents error when orchestration client is not set
+var ErrNotSet = errors.New("orchestration client not set")
+
+type Node interface {
+	Name() string
+	Client() *bee.Client
+	ClefKey() string
+	ClefPassword() string
+	Config() *Config
+	Create(ctx context.Context, o CreateOptions) (err error)
+	Delete(ctx context.Context, namespace string) (err error)
+	LibP2PKey() string
+	Ready(ctx context.Context, namespace string) (ready bool, err error)
+	Start(ctx context.Context, namespace string) (err error)
+	Stop(ctx context.Context, namespace string) (err error)
+	SwarmKey() string
 }
 
 // NodeOptions holds optional parameters for the Node.
@@ -21,68 +32,93 @@ type NodeOptions struct {
 	ClefKey      string
 	ClefPassword string
 	Client       *bee.Client
-	Config       *ob.Config
+	Config       *Config
+	K8S          *k8s.Client
 	LibP2PKey    string
 	SwarmKey     string
 }
 
-// NewNode returns Bee node
-func NewNode(name string, opts NodeOptions) (n *Node) {
-	n = &Node{name: name}
-
-	if opts.Client != nil {
-		n.client = opts.Client
-	}
-	if opts.Config != nil {
-		n.config = opts.Config
-	}
-	if len(opts.ClefKey) > 0 {
-		n.clefKey = opts.ClefKey
-	}
-	if len(opts.ClefPassword) > 0 {
-		n.clefPassword = opts.ClefPassword
-	}
-	if len(opts.LibP2PKey) > 0 {
-		n.libP2PKey = opts.LibP2PKey
-	}
-	if len(opts.SwarmKey) > 0 {
-		n.swarmKey = opts.SwarmKey
-	}
-
-	return
+// CreateOptions represents available options for creating node
+type CreateOptions struct {
+	// Bee configuration
+	Config Config
+	// Kubernetes configuration
+	Name                      string
+	Namespace                 string
+	Annotations               map[string]string
+	ClefImage                 string
+	ClefImagePullPolicy       string
+	ClefKey                   string
+	ClefPassword              string
+	Labels                    map[string]string
+	Image                     string
+	ImagePullPolicy           string
+	ImagePullSecrets          []string
+	IngressAnnotations        map[string]string
+	IngressClass              string
+	IngressHost               string
+	IngressDebugAnnotations   map[string]string
+	IngressDebugClass         string
+	IngressDebugHost          string
+	LibP2PKey                 string
+	NodeSelector              map[string]string
+	PersistenceEnabled        bool
+	PersistenceStorageClass   string
+	PersistenceStorageRequest string
+	PodManagementPolicy       string
+	RestartPolicy             string
+	ResourcesLimitCPU         string
+	ResourcesLimitMemory      string
+	ResourcesRequestCPU       string
+	ResourcesRequestMemory    string
+	Selector                  map[string]string
+	SwarmKey                  string
+	UpdateStrategy            string
 }
 
-// Name returns node's name
-func (n *Node) Name() string {
-	return n.name
-}
-
-// Client returns node's name
-func (n *Node) Client() *bee.Client {
-	return n.client
-}
-
-// Config returns node's config
-func (n *Node) Config() *ob.Config {
-	return n.config
-}
-
-// ClefKey returns node's clefKey
-func (n *Node) ClefKey() string {
-	return n.clefKey
-}
-
-// ClefPassword returns node's clefPassword
-func (n *Node) ClefPassword() string {
-	return n.clefPassword
-}
-
-// LibP2PKey returns node's libP2PKey
-func (n *Node) LibP2PKey() string {
-	return n.libP2PKey
-}
-
-// SwarmKey returns node's swarmKey
-func (n *Node) SwarmKey() string {
-	return n.swarmKey
+// Config represents Bee configuration
+type Config struct {
+	APIAddr                    string        // HTTP API listen address
+	BlockTime                  uint64        // chain block time
+	Bootnodes                  string        // initial nodes to connect to
+	BootnodeMode               bool          // cause the node to always accept incoming connections
+	CacheCapacity              uint64        // cache capacity in chunks, multiply by 4096 (MaxChunkSize) to get approximate capacity in bytes
+	ClefSignerEnable           bool          // enable clef signer
+	ClefSignerEndpoint         string        // clef signer endpoint
+	CORSAllowedOrigins         string        // origins with CORS headers enabled
+	DataDir                    string        // data directory
+	DbOpenFilesLimit           int           // number of open files allowed by database
+	DbBlockCacheCapacity       int           // size of block cache of the database in bytes
+	DbWriteBufferSize          int           // size of the database write buffer in bytes
+	DbDisableSeeksCompaction   bool          // disables DB compactions triggered by seeks
+	DebugAPIAddr               string        // debug HTTP API listen address
+	DebugAPIEnable             bool          // enable debug HTTP API
+	FullNode                   bool          // cause the node to start in full mode
+	GatewayMode                bool          // disable a set of sensitive features in the api
+	GlobalPinningEnabled       bool          // enable global pinning
+	NATAddr                    string        // NAT exposed address
+	NetworkID                  uint64        // ID of the Swarm network
+	P2PAddr                    string        // P2P listen address
+	P2PQUICEnable              bool          // enable P2P QUIC transport
+	P2PWSEnable                bool          // enable P2P WebSocket transport
+	Password                   string        // password for decrypting keys
+	PaymentEarly               uint64        // amount in BZZ below the peers payment threshold when we initiate settlement
+	PaymentThreshold           uint64        // threshold in BZZ where you expect to get paid from your peers
+	PaymentTolerance           uint64        // excess debt above payment threshold in BZZ where you disconnect from your peer
+	PostageStampAddress        string        // postage stamp address
+	PriceOracleAddress         string        // price Oracle address
+	ResolverOptions            string        // ENS compatible API endpoint for a TLD and with contract address, can be repeated, format [tld:][contract-addr@]url
+	Standalone                 bool          // whether we want the node to start with no listen addresses for p2p
+	SwapEnable                 bool          // enable swap
+	SwapEndpoint               string        // swap ethereum blockchain endpoint
+	SwapDeploymentGasPrice     string        // gas price in wei to use for deployment and funding
+	SwapFactoryAddress         string        // swap factory address
+	SwapLegacyFactoryAddresses string        // swap legacy factory addresses
+	SwapInitialDeposit         uint64        // initial deposit if deploying a new chequebook
+	TracingEnabled             bool          // enable tracing
+	TracingEndpoint            string        // endpoint to send tracing data
+	TracingServiceName         string        // service name identifier for tracing
+	Verbosity                  uint64        // log verbosity level 0=silent, 1=error, 2=warn, 3=info, 4=debug, 5=trace
+	WelcomeMessage             string        // send a welcome message string during handshakes
+	WarmupTime                 time.Duration // warmup time pull/pushsync protocols
 }
