+++ pkg/simulate/upload/upload.go
@@ -61,7 +61,7 @@ func NewSimulation() beekeeper.Action {
 }
 
 // Run executes upload stress
-func (s *Simulation) Run(ctx context.Context, cluster *orchestration.Cluster, opts interface{}) (err error) {
+func (s *Simulation) Run(ctx context.Context, cluster orchestration.Cluster, opts interface{}) (err error) {
 	fmt.Println("running upload simulation")
 	o, ok := opts.(Options)
 	if !ok {
