+++ pkg/test/bee.go
@@ -18,7 +18,7 @@ type CheckCase struct {
 	ctx      context.Context
 	clients  map[string]*bee.Client
 	nodes    []nodeV2
-	cluster  *orchestration.Cluster
+	cluster  orchestration.Cluster
 	overlays orchestration.ClusterOverlays
 
 	o               CaseOptions
@@ -37,7 +37,7 @@ type CaseOptions struct {
 	PostageDepth  uint64
 }
 
-func NewCheckCase(ctx context.Context, cluster *orchestration.Cluster, o CaseOptions) (*CheckCase, error) {
+func NewCheckCase(ctx context.Context, cluster orchestration.Cluster, o CaseOptions) (*CheckCase, error) {
 	clients, err := cluster.NodesClients(ctx)
 	if err != nil {
 		return nil, err
